+++ lib/facil/fiobj/fiobj_ary.h
@@ -53,14 +53,14 @@ FIOBJ *fiobj_ary2ptr(FIOBJ ary);
  * Negative values are retrieved from the end of the array. i.e., `-1`
  * is the last item.
  */
-FIOBJ fiobj_ary_index(FIOBJ ary, int64_t pos);
+FIOBJ fiobj_ary_index(FIOBJ ary, int32_t pos);
 /** alias for `fiobj_ary_index` */
 #define fiobj_ary_entry(a, p) fiobj_ary_index((a), (p))
 
 /**
  * Sets an object at the requested position.
  */
-void fiobj_ary_set(FIOBJ ary, FIOBJ obj, int64_t pos);
+void fiobj_ary_set(FIOBJ ary, FIOBJ obj, int32_t pos);
 
 /* *****************************************************************************
 Array push / shift API
@@ -91,13 +91,13 @@ Array Find / Remove / Replace
  * Replaces the object at a specific position, returning the old object -
  * remember to `fiobj_free` the old object.
  */
-FIOBJ fiobj_ary_replace(FIOBJ ary, FIOBJ obj, int64_t pos);
+FIOBJ fiobj_ary_replace(FIOBJ ary, FIOBJ obj, int32_t pos);
 
 /**
  * Finds the index of a specifide object (if any). Returns -1 if the object
  * isn't found.
  */
-int64_t fiobj_ary_find(FIOBJ ary, FIOBJ data);
+int32_t fiobj_ary_find(FIOBJ ary, FIOBJ data);
 
 /**
  * Removes the object at the index (if valid), changing the index of any
@@ -105,7 +105,7 @@ int64_t fiobj_ary_find(FIOBJ ary, FIOBJ data);
  *
  * Returns 0 on success or -1 (if no object or out of bounds).
  */
-int fiobj_ary_remove(FIOBJ ary, int64_t pos);
+int fiobj_ary_remove(FIOBJ ary, int32_t pos);
 
 /**
  * Removes any instance of an object from the Array (if any), changing the
